+++ Vault-Coder/src/utopia/vault/coder/controller/writer/ModelWriter.scala
@@ -67,6 +67,10 @@ object ModelWriter
 					Parameter("id", idType, description = s"id of this ${classToWrite.name} in the database"),
 					Parameter("data", dataClassRef, description = s"Wrapped ${classToWrite.name} data")
 				)
+				val singleAccessRef = AccessWriter.singleIdReferenceFor(classToWrite)
+				val accessProperty = ComputedProperty("access", Set(singleAccessRef),
+					description = s"An access point to this ${classToWrite.name} in the database")(
+					s"${singleAccessRef.target}(id)")
 				val description = s"Represents a ${classToWrite.name} that has already been stored in the database"
 				// ModelConvertible extension & implementation differs based on id type
 				if (classToWrite.useLongId)
@@ -74,11 +78,13 @@ object ModelWriter
 						Vector(Reference.stored(dataClassRef, idType)),
 						properties = Vector(
 							ComputedProperty("toModel", Set(Reference.valueConversions, Reference.constant),
-								isOverridden = true)("Constant(\"id\", id) + data.toModel")
+								isOverridden = true)("Constant(\"id\", id) + data.toModel"),
+							accessProperty
 						), description = description, isCaseClass = true)
 				else
 					declaration.ClassDeclaration(classToWrite.name.singular, constructionParams,
 						Vector(Reference.storedModelConvertible(dataClassRef)),
+						properties = Vector(accessProperty),
 						description = description, author = classToWrite.author, isCaseClass = true)
 			}
 			File(storePackage, storedClass).write().map { _ -> dataClassRef }